Toggle driving method and speed
The button on the side of the joystick now toggles between point and Z-Axis and the trigger now toggles between full and half speed in those same driving methods. There is an indicator on the SmartDashboard that tells when the speed is toggled. The last thing for programming to do is to add some kind of indication about whether or not the driving method is toggled, whether by dynamically updating the radio button or having another text box that states whether or not it has been toggled.